// src/types/LanguageTags.ts
export type Tags =
  | "C"
  | "C++"
  | "Java"
  | "JavaScript"
  | "TypeScript"
  | "Python"
  | "Node.js"
  | "HTML"
  | "CSS"
  | "React"
  | "Next.js"
  | "TailwindCSS"
  | "Framer Motion"
  | "Figma"
  | "Express"
  | "Recoil"
  | "WebSockets"
  | "MongoDB"
  | "PostgreSQL"
  | "MySQL"
  | "Prisma"
  | "Git"
  | "GitHub"
  | "Vercel"
  | "Google OAuth 2.0"
  | "NextAuth.js"
  | "REST APIs"
  | "OneStop";